(Doc+) Add note for `regexes.yaml` location
Hi team,

SF: 01833308 

I'd like to request adding a note to the documentation regarding the `regexes.yaml` file location. Currently, there is no information about where to find this file. Users should be aware that the `regexes.yaml` file is not directly accessible in the filesystem and is bundled within the useragent plugin JAR file.

Please consider adding this clarification to the documentation section.